requirements.txt
pre-commit >= 3.0.4, <4.0.0
pydantic >= 1.10.4, <2.0
python-dateutil >= 2.8.2, <3.0.0
requests >= 2.28.2, <3.0
